WebOfScience module - initial attempt WIP
