+++ spec/lib/web_of_science/queries_spec.rb
@@ -2,7 +2,7 @@
 # require the helper module
 require 'savon/mock/spec_helper'
 
-describe WosQueries do
+describe WebOfScience::Queries do
   include Savon::SpecHelper
 
   # set Savon in and out of mock mode
